+++ src/main/java/es/deusto/spq/server/data/jdo/User.java
@@ -1,4 +1,4 @@
-package es.deusto.spq.server.data.dao;
+package es.deusto.spq.server.data.jdo;
 
 
 import javax.jdo.annotations.*;
